Fix removed np.int alias in nstates return annotation
np.int was removed in NumPy 1.24, and the project pins numpy<2, so no
supported NumPy version defines it. nstates returns len(self.electronic_states),
a builtin int, so the annotation was also semantically wrong. It is currently
masked only because 'from __future__ import annotations' makes annotations lazy
strings, but typing.get_type_hints() on this class would raise AttributeError.
This is the only such deprecated alias left in the codebase.

+++ mlatom/data.py
@@ -1322,7 +1322,7 @@ class molecule:
 
 
     @property 
-    def nstates(self) -> np.int:
+    def nstates(self) -> int:
         '''
         The number of electronic states.
         '''
